feat(translate): integrate TranslateMenuProvider and refactor translation handling
